Add support for intended recipient subpackets on signatures
New API for hiding and checking recipients in message encryption
Adds the intended recipient subpacket in signatures. If a message is
encrypted and signed it now adds the intended recipients to the signature
(if not hidden). If the signature contains inteded recipients during a read
and the config flag is set, it checks if the primary id of the decryption key is
present in the list.