Add function DetachSignWriter for detached singatures
This commit unifiies the API for detached signatures with the API for encrypting messages.
The DetachSign function requires the caller to provide a reader for the message, while
encrypt returns a writecloser to which the message is wirtten to.
To unifiy the API and allow combinations, this commit adds a function DetachSignWriter
which returns a writecloser similar to the encryption API.