Ensure that intended recipients are checked
This commit changes the behaviour of the intended recipients check on
unencrypted messages. GopenPGP requires to call this function on a decrypted message,
which triggered that the recipients are not checked in the signature even if the
intended recipients check flag is set to true.
Now it does check it if the flag is true.