Allow multiple signatures while reading PGP messages
For the transition to newer versions of the OpenPGP standard, it is important
to support multiple signatures in messages. For example, a message might be
signed with a v4, a v6, and a vX key for compatibility. This commit modifies the MessageDetails
struct to track multiple signatures at the same time instead of one (Not backward compatible).
If one of these signature candidates can be successfully verified, the message is considered valid.
Currently, there is no priority, but a priority for the newest version could be implemented.