Allow multiple signatures while writing PGP messages
For the transition to newer versions of the OpenPGP standard, it is important
to support multiple signatures in messages. For example, a message might be
signed with a v4, a v6, and a vX key for compatibility. This commit modifies all signing
APIs to support more than one signing key, and produces a signature for each provided entity.