Ensure that critical unknown packet tags are rejected
The crypto-refresh mandets that If an implementation encounters a critical packet where the packet type
is unknown in a packet sequence, it MUST reject the whole packet sequence.
Packet Tags from 0 to 39 are critical. See:
https://www.ietf.org/archive/id/draft-ietf-openpgp-crypto-refresh-08.html#name-packet-criticality

The packet.Read method now returns a CriticalUnknownPacketTypeError if a critical unknown packet is detectded
and the packet.Reader forwards this error.
To avoid isses unknown critical packets, this commit additionaly adds a marker packet and trust packet type.