Rework key validation and signature verification
This commit changes how and when signatures are verified in keys (i.e., key life cycle).
Instead of verifing key signatures at parsing time, keys are now validate when used
similar to the key handling in OpenPGP.js.
When verifying a signature in a pgp message, the library now verifies that the key is
valid at signature ceation time.