Make entitiy parsing more robust to unknown keys
Previous to this commit, the entitiy parse method fails if the keyring contains
an unsupoorted subkey. It fails because an unkown public key packet is skipped while
the following signature packets are not. This commit introduces an new method on the
packet parser called NextWithUnsupported that signals to the caller that an unsuported packet
was parsed. This method allows to introduce a new check in entity parse to skip signatures from
unsupported keys and make the whole keyring parseable.